Optimize Toast related code
